+++ src/components/Card/index.jsx
@@ -1,4 +1,4 @@
-import "./Card.css";
+import "./style.css";
 import { getDate, getElapsedTime } from "../../utils";
 
 export default function Card({ data }) {
