+++ src/components/Footer/index.jsx
@@ -1,4 +1,4 @@
-import "./Footer.css";
+import "./style.css";
 
 export default function Footer() {
   return (
